requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-appconfiguration
azure-cosmosdb-table
azure-servicebus
requests
beautifulsoup4
twilio
croniter
opencensus-ext-azure
opencensus-ext-logging